Refactor Emacs configurations and update pre-commit hooks
Removed the end-of-file-fixer hook from the pre-commit
configuration. Updated the Emacs README with improved eglot server
program settings for `nix-mode` and `markdown-mode`. Added eglot
support for `yaml-ts-mode` and `toml-ts-mode`, ensuring the respective
language servers are set correctly. Introduced a new entry for
`yaml-language-server` in the packages.toml file for better language
support.